Refactor path resolution and use `@/` for vault root.
